refactor: /reservation-mine endpoint /reservation/mine 로 수정
