+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS reservation_time
 (
-    id       BIGINT       NOT NULL AUTO_INCREMENT,
+    id       BIGINT       NOT NULL,
 --     TODO: UNIQUE 속성 넣을지 고민하기
     start_at VARCHAR(255) NOT NULL,
     PRIMARY KEY (id)
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS reservation_time
 
 CREATE TABLE IF NOT EXISTS theme
 (
-    id          BIGINT       NOT NULL AUTO_INCREMENT,
+    id          BIGINT       NOT NULL,
     name        VARCHAR(255) NOT NULL,
     description VARCHAR(255) NOT NULL,
     thumbnail   VARCHAR(255) NOT NULL,
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS theme
 
 CREATE TABLE IF NOT EXISTS member
 (
-    id       BIGINT       NOT NULL AUTO_INCREMENT,
+    id       BIGINT       NOT NULL,
     name     VARCHAR(255) NOT NULL,
     email    VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS member
 
 CREATE TABLE IF NOT EXISTS reservation
 (
-    id        BIGINT       NOT NULL AUTO_INCREMENT,
+    id        BIGINT       NOT NULL,
     date      VARCHAR(255) NOT NULL,
     time_id   BIGINT,
     theme_id  BIGINT,
