+++ src/main/resources/templates/reservation.html
@@ -43,7 +43,7 @@
           <span id="profile-name">Profile</span> <!-- 프로필 이름을 넣을 span 추가 -->
         </a>
         <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-          <a class="dropdown-item" href="/reservation-mine">My Reservation</a>
+          <a class="dropdown-item" href="/reservation/mine">My Reservation</a>
           <div class="dropdown-divider"></div>
           <a class="dropdown-item" href="#" id="logout-btn">Logout</a>
         </div>
